# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: GSI Partner Opportunity Tracker API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all users (admin only)
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      operationId: createUser
      tags: [users]
      summary: Create a new user (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserInput"
      responses:
        "201":
          description: Created user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get user by id
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: Not found
    patch:
      operationId: updateUser
      tags: [users]
      summary: Update user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    delete:
      operationId: deleteUser
      tags: [users]
      summary: Delete user (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # PROFILE
...
  /reminders:
    get:
      operationId: listReminders
      tags: [reminders]
      summary: List reminders (sorted by due date)
      parameters:
        - name: status
          in: query
          description: Filter by completion status (default all)
          schema:
            type: string
            enum: [open, completed, all]
      responses:
        "200":
          description: List of reminders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Reminder"
    post:
      operationId: createReminder
      tags: [reminders]
      summary: Create a reminder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReminderInput"
      responses:
        "201":
          description: Created reminder
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reminder"
  /reminders/{id}:
    patch:
      operationId: updateReminder
      tags: [reminders]
      summary: Update a reminder (rename, notes, reschedule, complete/un-complete)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReminderUpdate"
      responses:
        "200":
          description: Updated reminder
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reminder"
        "404":
          description: Not found
    delete:
      operationId: deleteReminder
      tags: [reminders]
      summary: Delete reminder
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # PARTNER RESOURCES
...
components:
  schemas:
...
    User:
      type: object
      required: [id, email, role, name]
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [admin, rep]
        quota:
          type: ["number", "null"]
        region:
          type: ["string", "null"]
        createdAt:
          type: string
    UserInput:
      type: object
      required: [email, password, name, role]
      properties:
        email:
          type: string
        password:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [admin, rep]
        quota:
          type: ["number", "null"]
        region:
          type: ["string", "null"]
    UserUpdate:
      type: object
      properties:
        name:
          type: string
        role:
          type: string
          enum: [admin, rep]
        quota:
          type: ["number", "null"]
        region:
          type: ["string", "null"]
        password:
          type: ["string", "null"]

    # PROFILE
...
    Reminder:
      type: object
      required: [id, name, dueAt, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        dueAt:
          type: string
        entityType:
          type: ["string", "null"]
        entityId:
          type: ["string", "null"]
        entityLabel:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        completedAt:
          type: ["string", "null"]
        createdAt:
          type: string
    ReminderInput:
      type: object
      required: [name, dueAt]
      properties:
        name:
          type: string
        dueAt:
          type: string
        entityType:
          type: ["string", "null"]
        entityId:
          type: ["string", "null"]
        entityLabel:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
    ReminderUpdate:
      type: object
      description: Partial update. Set completed=true/false to complete or reopen.
      properties:
        name:
          type: string
        dueAt:
          type: string
        notes:
          type: ["string", "null"]
        completed:
          type: boolean
        entityType:
          type: ["string", "null"]
        entityId:
          type: ["string", "null"]
        entityLabel:
          type: ["string", "null"]
...