# lib/api-zod/src/generated/types/reminder.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * GSI Partner Opportunity Tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface Reminder {
  id: number;
  name: string;
  dueAt: string;
  /** @nullable */
  entityType?: string | null;
  /** @nullable */
  entityId?: string | null;
  /** @nullable */
  entityLabel?: string | null;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  completedAt?: string | null;
  createdAt: string;
}